requirements.txt
python-dotenv
pyupbit
grpcio==1.67.1
requests
google-generativeai
ta